% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{brf_list_ticker_roots}
\alias{brf_list_ticker_roots}
\title{Retrieve available ticker roots from the B3 bulletin}
\usage{
brf_list_ticker_roots()
}
\value{
A data frame with `ticker_root` and `description` columns.
}
\description{
Retrieve available ticker roots from the B3 bulletin
}
