% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brfutures-package.R
\docType{package}
\name{brfutures-package}
\alias{brfutures}
\alias{brfutures-package}
\title{brfutures: Brazilian Futures Data Tools}
\description{
Helpers to download, parse, and organise Brazilian futures bulletins from B3 (formerly BM&F) with ready-to-use xts time series.
}
\author{
\strong{Maintainer}: Hugo Rzepian Teixeira \email{hugorteixeira@senhormercado.com.br}

}
\keyword{internal}
